+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- fix(android): setContext serializes as context for Android instead of extra (#2155)
 - fix(android): Duplicate Breadcrumbs when captuing messages #2153
 
 ## 3.3.3